split out ip discoverey and improve it from a bunch of punning
